+++ rust/main/lander/src/adapter/chains/ethereum/adapter/gas_price/escalator.rs
@@ -49,6 +49,7 @@ pub fn escalate_gas_price_if_needed(
             debug!(
                 tx_type = "Legacy or Eip2930",
                 ?old_gas_price,
+                ?estimated_gas_price,
                 ?escalated_gas_price,
                 "Escalation attempt outcome"
             );
@@ -63,21 +64,25 @@ pub fn escalate_gas_price_if_needed(
                 max_priority_fee: old_max_priority_fee,
             },
             GasPrice::Eip1559 {
-                max_fee: new_max_fee,
-                max_priority_fee: new_max_priority_fee,
+                max_fee: estimated_max_fee,
+                max_priority_fee: estimated_max_priority_fee,
             },
         ) => {
             let escalated_max_fee_per_gas =
-                get_escalated_price_from_old_and_new(old_max_fee, new_max_fee);
+                get_escalated_price_from_old_and_new(old_max_fee, estimated_max_fee);
 
-            let escalated_max_priority_fee_per_gas =
-                get_escalated_price_from_old_and_new(old_max_priority_fee, new_max_priority_fee);
+            let escalated_max_priority_fee_per_gas = get_escalated_price_from_old_and_new(
+                old_max_priority_fee,
+                estimated_max_priority_fee,
+            );
 
             debug!(
                 tx_type = "Eip1559",
                 old_max_fee_per_gas = ?old_max_fee,
+                estimated_max_fee = ?estimated_max_fee,
                 escalated_max_fee_per_gas = ?escalated_max_fee_per_gas,
                 old_max_priority_fee_per_gas = ?old_max_priority_fee,
+                estimated_max_priority_fee = ?estimated_max_priority_fee,
                 escalated_max_priority_fee_per_gas = ?escalated_max_priority_fee_per_gas,
                 "Escalation attempt outcome"
             );
@@ -102,5 +107,35 @@ fn get_escalated_price_from_old_and_new(old_gas_price: &U256, new_gas_price: &U2
 fn apply_escalation_multiplier(gas_price: &U256) -> U256 {
     let numerator = U256::from(ESCALATION_MULTIPLIER_NUMERATOR);
     let denominator = U256::from(ESCALATION_MULTIPLIER_DENOMINATOR);
-    gas_price * numerator / denominator
+    gas_price.saturating_mul(numerator).div_mod(denominator).0
+}
+
+#[cfg(test)]
+mod tests {
+    use hyperlane_core::U256;
+
+    use crate::adapter::chains::ethereum::adapter::gas_price::GasPrice;
+
+    use super::*;
+
+    #[test]
+    fn test_gas_price_does_not_overflow() {
+        let old_gas_price = GasPrice::Eip1559 {
+            max_fee: U256::MAX,
+            max_priority_fee: U256::MAX,
+        };
+        let estimated_gas_price = GasPrice::Eip1559 {
+            max_fee: U256::MAX,
+            max_priority_fee: U256::MAX,
+        };
+
+        // should not overflow and panic
+        let res = escalate_gas_price_if_needed(&old_gas_price, &estimated_gas_price);
+        let expected = GasPrice::Eip1559 {
+            max_fee: U256::MAX,
+            max_priority_fee: U256::MAX,
+        };
+
+        assert_eq!(res, expected);
+    }
 }